rename device and consumer
